backend/requirements.txt
azure-ai-documentintelligence>=1.0.0
azure-core>=1.29.0
python-dotenv>=1.0.0
openai>=1.0.0
requests>=2.31.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6